Separate send component
